move configurations to test fixtures for downstream use in the REST API (#956)
Refactor db configuration classes to the fixtures packages.
