+++ integrationtesting/fixtures/src/main/java/org/opendcs/fixtures/configurations/opendcs/oracle/OpenDCSOracleConfigurationProvider.java
@@ -2,8 +2,8 @@ package org.opendcs.fixtures.configurations.opendcs.oracle;
 
 import java.io.File;
 
-import org.opendcs.spi.configuration.Configuration;
-import org.opendcs.spi.configuration.ConfigurationProvider;
+import org.opendcs.fixtures.spi.Configuration;
+import org.opendcs.fixtures.spi.ConfigurationProvider;
 
 public class OpenDCSOracleConfigurationProvider implements ConfigurationProvider
 {
