+++ integrationtesting/fixtures/src/main/java/org/opendcs/fixtures/configurations/xml/XmlConfiguration.java
@@ -11,15 +11,11 @@ import org.apache.commons.io.FileUtils;
 import org.opendcs.database.DatabaseService;
 import org.opendcs.database.api.OpenDcsDatabase;
 import org.opendcs.fixtures.UserPropertiesBuilder;
-import org.opendcs.spi.configuration.Configuration;
+import org.opendcs.fixtures.spi.Configuration;
 
 import decodes.db.Database;
-import decodes.db.DatabaseIO;
 import decodes.launcher.Profile;
-import decodes.tsdb.TimeSeriesDb;
 import decodes.util.DecodesSettings;
-import ilex.util.Pair;
-import decodes.xml.CreatePlatformXref;
 import uk.org.webcompere.systemstubs.environment.EnvironmentVariables;
 import uk.org.webcompere.systemstubs.properties.SystemProperties;
 import uk.org.webcompere.systemstubs.security.SystemExit;
