+++ integrationtesting/fixtures/src/main/resources/META-INF/services/org.opendcs.fixtures.spi.ConfigurationProvider
@@ -1,0 +1,4 @@
+org.opendcs.fixtures.configurations.xml.XmlConfigurationProvider
+org.opendcs.fixtures.configurations.opendcs.pg.OpenDCSPGConfigurationProvider
+org.opendcs.fixtures.configurations.cwms.CwmsOracleConfigurationProvider
+org.opendcs.fixtures.configurations.opendcs.oracle.OpenDCSOracleConfigurationProvider
