+++ integrationtesting/opendcs-tests/build.gradle
@@ -9,11 +9,6 @@ dependencies {
     testCompileOnly project(":opendcs")
     testImplementation project(":testing:fixtures")
     testImplementation enforcedPlatform(libs.junit.bom)
-    testImplementation(libs.testcontainers.core)
-    testImplementation(libs.testcontainers.postgresql)
-    testImplementation(libs.testcontainers.oracle.free)
-    testImplementation(libs.hec.testcontainers.cwms)
-    testImplementation(libs.bundles.jdbi)
     testImplementation(libs.bundles.junit)
     testImplementation(libs.bundles.junit.platform)
     testImplementation(libs.commons.io)
