--- integrationtesting/opendcs-tests/src/test/resources/META-INF/services/org.opendcs.spi.configuration.ConfigurationProvider
@@ -1,4 +1,0 @@
-org.opendcs.fixtures.configurations.xml.XmlConfigurationProvider
-org.opendcs.fixtures.configurations.opendcs.pg.OpenDCSPGConfigurationProvider
-org.opendcs.fixtures.configurations.opendcs.oracle.OpenDCSOracleConfigurationProvider
-org.opendcs.fixtures.configurations.cwms.CwmsOracleConfigurationProvider
