+++ java/annotations/src/main/java/org/opendcs/annotations/algorithm/Output.java
@@ -25,8 +25,7 @@ import java.lang.annotation.Target;
 public @interface Output
 {
     /**
-     * Leaveing the name blank will use the name of the field.
-     * @return
+     * Leaving the name blank will use the name of the field.
      */
     String name() default "";
     Class<?> type() default Double.class;
